@@ -1,18 +1,18 @@
 {
   "name": "@optimumfuturist/scrum",
   "version": "3.0.0",
-  "homepage": "./",
+  "homepage": "./scrum",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^13.1.9",
-    "localforage": "^1.9.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "localforage": "^1.10.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-impressjs": "^1.1.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "uuid": "^8.3.2"
+    "react-router-dom": "^6.8.2",
+    "react-scripts": "^5.0.1",
+    "uuid": "^9.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
