@@ -1,9 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
+import localforage from 'localforage';
+
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import localforage from 'localforage';
 import { LF_KEYS } from './constants';
 
 localforage.config({
@@ -19,12 +20,10 @@ localforage.getItem(LF_KEYS.PERSONS).then(values => {
 });
 
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render(<App />);
+
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
